+++ components/chart/yAxis.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import { LegislatureType } from "@/types/legislature";
 import { motion } from "framer-motion";
-import getDate from "../utils/getDate";
+import getDate from "@/utils/getDate";
 import { scaleLinear } from "d3";
 
 // Inspired by https://2019.wattenberger.com/blog/react-and-d3
